+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## dev
+
+* Bug
+  * [Core] Fixing regex bug in expand envs (not expand numbers: `$1` or `${2}`);
+  * [Core] Adding image envs in expand envs variable process;
+  * [Core] Adding support to escape variables in command options (shell and system);
+
 ## v0.18.0 - (2016-04-07)
 
 * Enhancements